# public/swagger/swagger.yaml
openapi: '3.0.0'
info:
  title: 'WeForYouth API'
  version: '1.0.0'
  description: 'WeForYouth REST API For WeForYouth App'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
definitions:        # Schema defination for request body
  users:
    type: object
    required:
      - name
      - email
      - password
      - age
      - gender
      - community
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      age:
        type: string
      gender:
        type: string
      community:
        type: string
  roles:
    type: object
    required:
      - rolename
      - user
    properties:
      rolename:
        type: string
      user:
        type: string

paths:
  /api/register:           # path of your endpoint
    post:
      tags:               # Tag property
        - User         # Value of the tag
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/users'
      responses:
        '200':
          description: User is successfully registered
        '500':
          description: Internal server error
        '403':
          description: Unauthorized
        '401':
          description: Unauthenticated
        '400':
          description: Bad request

  /api/user/login:
    post:
      tags:
        - User
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                - email:
                    type: string
                    description: Email Address of the user
                - password:
                    type: string
                    description: Password of the user
            example:
              email: email
              password: password
      responses:
        '200':
          description: Succefully Assigned Manager To Accomodation
        '500':
          description: Internal server error
        '404':
          description: Not Found
        '401':
          description: Unauthenticated
        '409':
          description: Conflict
        '411':
          description: Invalid Credentials
        '422':
          description: Invalid data in the request

  /api/role/assignRole:
    put:
      tags:
        - Roles
      summary: Assign a user a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/roles'
      responses:
        '200':
          description: Role status changed
        '500':
          description: Internal server error
        '404':
          description: Not Found
        '401':
          description: Unauthenticated/Not authorized
        '409':
          description: Conflict
        '422':
          description: Invalid data in the request